+++ src/main/java/de/officeryoda/fhysics/engine/FhysicsCore.kt
@@ -3,9 +3,7 @@ package de.officeryoda.fhysics.engine
 import de.officeryoda.fhysics.engine.collision.CollisionInfo
 import de.officeryoda.fhysics.engine.collision.CollisionSolver
 import de.officeryoda.fhysics.engine.collision.ElasticCollision
-import de.officeryoda.fhysics.engine.objects.Circle
-import de.officeryoda.fhysics.engine.objects.FhysicsObject
-import de.officeryoda.fhysics.engine.objects.Polygon
+import de.officeryoda.fhysics.engine.objects.*
 import de.officeryoda.fhysics.extensions.times
 import de.officeryoda.fhysics.rendering.FhysicsObjectDrawer
 import de.officeryoda.fhysics.rendering.GravityType
@@ -48,19 +46,20 @@ object FhysicsCore {
     init {
 
 
-//        for (i in 1..20) {
-//            val circle: Circle = FhysicsObjectFactory.randomCircle()
-////            circle.velocity.set(Vector2.ZERO)
-//            spawn(circle)
-//        }
-//
-//        for (i in 1..20) {
-//            val rect: Rectangle = FhysicsObjectFactory.randomRectangle()
-//            spawn(rect)
-//        }
-//
-//        for (i in 1..10)
-//            spawn(FhysicsObjectFactory.randomConvexPolygon())
+        for (i in 1..20) {
+            val circle: Circle = FhysicsObjectFactory.randomCircle()
+//            circle.velocity.set(Vector2.ZERO)
+            spawn(circle)
+        }
+
+        for (i in 1..20) {
+            val rect: Rectangle = FhysicsObjectFactory.randomRectangle()
+            spawn(rect)
+        }
+
+        for (i in 1..10) {
+            spawn(FhysicsObjectFactory.randomPolygon())
+        }
 
         // spawn a rotated rectangle in the center
 //        val rect =
