+++ src/main/java/de/officeryoda/fhysics/engine/objects/ConcavePolygon.kt
@@ -34,29 +34,3 @@ class ConcavePolygon(
     }
 }
 
-class SubPolygon(
-    position: Vector2,
-    center: Vector2,
-    velocity: Vector2,
-    vertices: Array<Vector2>,
-    rotation: Float = 0f,
-) : Polygon(position, velocity, vertices, rotation) {
-
-    private val centerOffset = center - position
-    override val center
-        get() = position + centerOffset
-
-
-    override fun testCollision(other: FhysicsObject): CollisionInfo {
-        return other.testCollision(this)
-    }
-
-    override fun clone(): FhysicsObject {
-        TODO()
-//        return SubPolygon(vertices.map { it.copy() }.toTypedArray(), indices, position.copy(), velocity.copy(), rotation)
-    }
-
-    override fun toString(): String {
-        return "SubPolygon(id=$id, position=$position, velocity=$velocity, acceleration=$acceleration, mass=$mass, static=$static, color=$color, vertices=${vertices.contentToString()})"
-    }
-}
