+++ src/main/java/de/officeryoda/fhysics/rendering/UIController.kt
@@ -502,7 +502,7 @@ class UIController {
         /// =====Gravity=====
         var gravityType: GravityType = GravityType.DIRECTIONAL
             private set
-        val gravityDirection: Vector2 = Vector2(0.0f, -10.0f)
+        val gravityDirection: Vector2 = Vector2(0.0f, 0.0f)
         val gravityPoint: Vector2 = Vector2( // The center of the world
             (FhysicsCore.BORDER.width / 2.0).toFloat(),
             (FhysicsCore.BORDER.height / 2.0).toFloat()
